commit
casi listo habitaciones

+++ CommonHotel/src/main/java/com/aa/common/dto/habitaciones/HabitacionResponse.java
@@ -1,10 +1,12 @@
 package com.aa.common.dto.habitaciones;
 
+import java.math.BigDecimal;
+
 public record HabitacionResponse(
 		Long idHabitacion,
 		Integer numeroHabitacion,
 		String tipoHabitacion,
-		Double precioNoche,
+		BigDecimal precioNoche,
 		Integer capacidad,
 		String estadoHabitacion,
 		String fechaCreacion,
